Revert change, old rust compat.
Rustc 1.25 and earlier does not have isize::from(u8), so use i8 for
byte calculations when parsing numbers.

+++ src/parser/value.rs
@@ -301,9 +301,11 @@ named!(
     map!(
         fold_many1!(
             // Note: We should use bytes directly, one_of returns a char.
+            // Also, rustc 1.25 and earlier does not have isize::from(u8),
+            // so use i8 for bytes.
             one_of!("0123456789"),
             0,
-            |r, d| r * 10 + isize::from(d as u8 - b'0')
+            |r, d| r * 10 + isize::from(d as i8 - b'0' as i8)
         ),
         Rational::from_integer
     )
@@ -315,7 +317,7 @@ named!(
         preceded!(
             complete!(tag!(".")),
             fold_many1!(one_of!("0123456789"), (0, 1), |(r, n), d| (
-                r * 10 + isize::from(d as u8 - b'0'),
+                r * 10 + isize::from(d as i8 - b'0' as i8),
                 n * 10
             ))
         ),
